The Rs: Rust, Ruby, Racket, and R
Outside of `red-lang`, Racket was the most difficult to get going out of
this bunch. It was in alpine's package repository, but my base image was
alpine:3.16 (and it entered alpine's package repository in 3.17).

To sort that out _slash, the repository responsible for the image all
of these are based off of, needed a little update.

`red-lang` was a whole different beast, because I'm being strict on
having each image have a similar feel by using the same base images,
this doesn't look to be an option because red is a 32-bit binary. One
day maybe, but not today.

86 to go.