Better Templates!
There are still 85 Hello World's to go, so 85 uses of a slightly better
way to initialize new Hello World's is definitely desireable! This
allows for:

`make new` (You get a new "template - date" folder)
`make new LANG=emojicode` (You get a new "emojicode" folder)

In addition to the new porceline, the template itself is also improved
to minimize the number of changes and files you need to modify (or
rename).

85 to go.