72 101 108 108 111 32 119 111 114 108 100!(x86-64)
Hello World, in x86-64 assembly. This was particularly interesting
because I'm on an ARM architecture. I don't want to force explicit
architectures for each of the languages, but I had to introduce
cross-architecture Dockerfile builds because this is explicitly x86
assembly -- so we need to be explicit about the platform we want to run
the container in.

This is done through inspecting the name of the make target and setting
the --platform option in the docker commands if x86_64 is present.

83 to go.