Erlang says Hello!
I've no written more than 0 lines of erlang! Erlang is on my list of
"maybe I should actually learn this language". Why? The fact that
WhatsApp had 50 engineers and a billion users (and used erlang) is
pretty impressive -- and I've always been a fan of RabbitMQ which is
written in Erlang.

72 to go.