Fix Prolog? (And quiet it's output)
As mysteries are, I would swear I was hitting a variety of issues with
prolog so I had to disable it. However, in attempting to replicate them
after a few days -- couldn't do it.

So let's see if CI passes for this, and if so, might as well merge it!

69 to go.