Github Actions! Better Emojis + Copy
Looking through the UI and a few things just didn't feel as awesome as
they could be -- if I can't have emojicode, I should at least have some
lovely emojis.

Also, Github recommends not running schedule actions "on-the-hour" and
to offset by a few minutes because of their load patterns.