C+R Say Hello!
There is a constraint on the character length of images in Dockerhub,
they must be at least 2 characters. Sadly, this causes issues with C and
R (and presumably up to 24 additional languages)

In addition -- this will test if the count of "Hello World"'s to go is
correct with GHA. So now, we don't need to put how many are to go in the
commit messages!

66 to go. ;)