Added Smalltalk example
This implementation pull what is, as of this writing, the most recent
stable version of GNU Smalltalk and builds it from source.

Of course, this isn't without it's challenges.  Because Alpine uses musl
and not glibc, the code doesn't immediately compile due to use of the
now-deprecated __va_copy function, so I've included a patch to replace
that usage with regular va_copy, which gets the code compiled in c89
mode for GCC 12.

As for the example, I made it slightly more elaborate than strictly
necessary.  I could've just called displayNl on a string, but what fun
is that?  Smalltalk is one of the original OO languages, so not
including an example that defines a class, instantiates it, and sends
it a message would be just silly, wouldn't it?