Actions to run on pull_request + push
The pull_request_target means "Run the actions from the PR target's
primary branch" -- which means any new changes aren't present. So
although that grants access to the secrets, it's not particularly useful
in this context.

So when we're not publishing the image, we don't log in. There is
no way to make secrets available to PRs (which prevents all sorts of
secret exfil problems) and not logging into dockerhub for this makes
sense.