Clean up build-test-publish action
This action has worked well, but has many language specific concepts in
it. Sadly, that makes the build-test-publish action less useable for
other purposes.

What are those other purposes you ask? Well, I'll be introducing the
base container for all of these containers into this repository -- and
it's going to be built and published in a slightly different way than
the existing containers (as it's not a language, it's the base
hello-world container).