Remove superfulous }
My poor aging eyes missed a bracket. This means my previous commit was
incorrect about the necessity of explicitly mentioning your
dependencies -- but I prefer that anyway.

Bash is my test-bed for these publishing changes in main. A better
approach would be to work off a branch and change the logic to run the
publishing from the branch while testing -- but instead of you get a
messy shotgun approach here in main.