Remove .base from composite dockerfile
This should never have snuck it's way into the composite dockerfile, as
it's not built along-side other images (the upstream version is always
used). This wasn't a problem before because the Dockerfile wasn't
parameterized, so it's existence in the composite dockerfile didn't
break the build.

In addition to this, because I tested my previous merge to main in a
branch -- and it passed all the builds, that problem build is included
in this change (which will validate the change).