Algol says Hello!
Algol took me on a journey! One of the more interesting stops was the
"Man or boy test" (literally the name of the wikipedia page). It's a
recursive function call with both referenced and copied stack frames
to push the compiler's recursive implementations to the extreme. Knuth
proposed it, I definitely don't understand it.

This implementation is not a man or a boy, although it'd be fantastic to
get that at least building/running.