Coffeescript says Hello!
Turns out, having an alpine-3.10 distro available is helpful for more
than just llvm8! node-npm isn't available in alpine-3.18, so we fall
back to our older-version to use the default installation instructions.
coffeescript transpiles into javascript and has the whole world of
javascript available to it.

Transpile is the world I forgot to use to describe the algol -> c
transpiler here!