Vlang says Hello!
Vlang is a vlang -> c transpiler, with a wonderful toolchain around
transpilation and compilation. There is a demonstration of translating
DOOM (yes... that DOOM) from C to V and building it (in under a
second!?). It can also transpile V to javascript, or WASM. For a 5-year
old language, this is impressive.

33 to go.