Ceylon says Hello!
The time-to-hello world for Ceylon was very high. I got caught up in the
community, asking "Is this a dead language?" without receiving responses
for years. https://github.com/eclipse-archived/ceylon/issues/7454

There are some passionate folks around Ceylon, but Red Hat kicked it
over to Eclipse -- and that's where it seems to have died. There are
some requests for community involvement, but it's fallen behind JVM
tooling.

32 to go.